Getting All WordPress Info
This is the prototype for grabbing an article's information.
All the possible required info is here except maybe the picture.
needs a more dynamic original article URL and needs to do it's work in the background without displaying over the screen like it does now. The prototype is working though.