Notifications Fully Working
I have implemented the class NotificationsHelper.java which does all the work to create and then make a notification. 
Also, there is a persistent service in this version that runs once per minute but is started in the main activity so the app must run once for it to start. We need to figure out how to make a service run on start.