Pull to manual refresh prototype
Got the pull down to manually refresh code working. The way the main page is set up it happens on the all bottom nav pages though.
This is the prototype so it writes to the log instead of doing database stuff which is as far as I could get right now.